# robocop: reduce string to symbol conversion for performance