--- a/lib/fluent/compat/socket_util.rb
+++ b/lib/fluent/compat/socket_util.rb
@@ -129,7 +129,7 @@ module Fluent
 
         def shutdown
           @loop.watchers.each { |w| w.detach }
-          @loop.stop if @loop.instance_variable_get("@running")
+          @loop.stop if @loop.instance_variable_get(:@running)
           @handler.close
           @thread.join