--- a/lib/fluent/plugin_id.rb
+++ b/lib/fluent/plugin_id.rb
@@ -57,13 +57,13 @@ module Fluent
     end
 
     def plugin_id_configured?
-      if instance_variable_defined?("@_id_configured")
+      if instance_variable_defined?(:@_id_configured)
         @_id_configured
       end
     end
 
     def plugin_id
-      if instance_variable_defined?("@id")
+      if instance_variable_defined?(:@id)
         @id || "object:#{object_id.to_s(16)}"
       else
         "object:#{object_id.to_s(16)}"