--- a/lib/fluent/system_config.rb
+++ b/lib/fluent/system_config.rb
@@ -186,13 +186,13 @@ module Fluent
         unless defined?($_system_config)
           $_system_config = nil
         end
-        (instance_variable_defined?("@_system_config") && @_system_config) ||
+        (instance_variable_defined?(:@_system_config) && @_system_config) ||
           $_system_config || Fluent::Engine.system_config
       end
 
       def system_config_override(opts={})
         require 'fluent/engine'
-        if !instance_variable_defined?("@_system_config") || @_system_config.nil?
+        if !instance_variable_defined?(:@_system_config) || @_system_config.nil?
           @_system_config = (defined?($_system_config) && $_system_config ? $_system_config : Fluent::Engine.system_config).dup
         end
         opts.each_pair do |key, value|