<!--
Thank you for contributing to Fluentd!
Your commits need to follow DCO: https://probot.github.io/apps/dco/
And please provide the following information to help us make the most of your pull request:
-->

**Which issue(s) this PR fixes**: 
Fixes #

**What this PR does / why we need it**: 

This is cosmetic change, it does not change behavior at all. It was detected by the following rubocop configuration:

```
  Performance/StringIdentifierArgument:
    Enabled: true
```

Apparently, string identifier argument should not be used.

Benchmark result:

```
  ruby 3.2.8 (2025-03-26 revision 13f495dc2c) +YJIT [x86_64-linux]
  Warming up --------------------------------------
  get instance variable with string notation
                         918.201k i/100ms
  get instance variable with symbol notation
                           2.186M i/100ms
  Calculating -------------------------------------
  get instance variable with string notation
                           10.480M (± 1.5%) i/s   (95.42 ns/i) -     53.256M in   5.082903s
  get instance variable with symbol notation
                           27.349M (± 1.3%) i/s   (36.56 ns/i) -    137.703M in   5.035946s

  Comparison:
  get instance variable with symbol notation: 27349072.3 i/s
  get instance variable with string notation: 10479751.9 i/s - 2.61x  slower

  ruby 3.2.8 (2025-03-26 revision 13f495dc2c) +YJIT [x86_64-linux]
  Warming up --------------------------------------
  defined? instance variable with string notation
                         919.834k i/100ms
  defined? instance variable with symbol notation
                           2.333M i/100ms
  Calculating -------------------------------------
  defined? instance variable with string notation
                            9.011M (± 3.6%) i/s  (110.98 ns/i) -     45.072M in   5.008998s
  defined? instance variable with symbol notation
                           27.433M (± 2.2%) i/s   (36.45 ns/i) -    137.670M in   5.021011s

  Comparison:
  defined? instance variable with symbol notation: 27432875.4 i/s
  defined? instance variable with string notation:  9010878.0 i/s - 3.04x  slower
```

Appendix: benchmark ruby script

```ruby
  require 'bundler/inline'
  gemfile do
    source 'https://rubygems.org'
    gem 'benchmark-ips'
    gem 'benchmark-memory'
  end

  class Dummy
    def initialize
      @running = nil
    end
  end

  Benchmark.ips do |x|
    @dummy = Dummy.new
    x.report("get instance variable with string notation") {
      @dummy.instance_variable_get("@running")
    }
    x.report("get instance variable with symbol notation") {
      @dummy.instance_variable_get(:@running)
    }
    x.compare!
  end

  Benchmark.ips do |x|
    @dummy = Dummy.new
    x.report("defined? instance variable with string notation") {
      @dummy.instance_variable_defined?("@running")
    }
    x.report("defined? instance variable with symbol notation") {
      @dummy.instance_variable_defined?(:@running)
    }
    x.compare!
  end
```

**Docs Changes**:

N/A

**Release Note**: 

N/A